requirements.txt
tensorflow
tensorflow-datasets
matplotlib
ipykernel
ipywidgets
keras
python-box
pyYAML
ensure
gdown==4.6.0

-e .